Remove 'issues' and rename 'pulls' to 'mungers'
