+++ mungegithub/mungegithub.go
@@ -23,8 +23,7 @@ import (
 	"time"
 
 	github_util "k8s.io/contrib/mungegithub/github"
-	"k8s.io/contrib/mungegithub/issues"
-	"k8s.io/contrib/mungegithub/pulls"
+	"k8s.io/contrib/mungegithub/mungers"
 
 	"github.com/golang/glog"
 	"github.com/spf13/cobra"
@@ -36,38 +35,30 @@ var (
 
 type mungeConfig struct {
 	github_util.Config
-	MinIssueNumber   int
-	IssueMungersList []string
-	PRMungersList    []string
-	Once             bool
-	Period           time.Duration
+	MinIssueNumber int
+	PRMungersList  []string
+	Once           bool
+	Period         time.Duration
 }
 
 func addMungeFlags(config *mungeConfig, cmd *cobra.Command) {
 	cmd.Flags().BoolVar(&config.Once, "once", false, "If true, run one loop and exit")
-	cmd.Flags().StringSliceVar(&config.IssueMungersList, "issue-mungers", []string{}, "A list of issue mungers to run")
 	cmd.Flags().StringSliceVar(&config.PRMungersList, "pr-mungers", []string{"blunderbuss", "lgtm-after-commit", "needs-rebase", "ok-to-test", "path-label", "ping-ci", "size", "submit-queue"}, "A list of pull request mungers to run")
 	cmd.Flags().DurationVar(&config.Period, "period", 30*time.Minute, "The period for running mungers")
 }
 
 func doMungers(config *mungeConfig) error {
-	if len(config.IssueMungersList) == 0 && len(config.PRMungersList) == 0 {
+	if len(config.PRMungersList) == 0 {
 		glog.Fatalf("must include at least one --issue-mungers or --pr-mungers")
 	}
 	for {
 		nextRunStartTime := time.Now().Add(config.Period)
-		if len(config.IssueMungersList) > 0 {
-			glog.Infof("Running issue mungers")
-			if err := issues.MungeIssues(&config.Config); err != nil {
-				glog.Errorf("Error munging issues: %v", err)
-			}
-		}
 		if len(config.PRMungersList) > 0 {
 			glog.Infof("Running PR mungers")
-			if err := pulls.EachLoop(&config.Config); err != nil {
+			if err := mungers.EachLoop(&config.Config); err != nil {
 				glog.Errorf("Error in EachLoop: %v", err)
 			}
-			if err := pulls.MungePullRequests(&config.Config); err != nil {
+			if err := mungers.MungePullRequests(&config.Config); err != nil {
 				glog.Errorf("Error munging PRs: %v", err)
 			}
 		}
@@ -95,21 +86,15 @@ func main() {
 			if err := config.PreExecute(); err != nil {
 				return err
 			}
-			issues.InitializeMungers(config.IssueMungersList, &config.Config)
-			pulls.InitializeMungers(config.PRMungersList, &config.Config)
+			mungers.InitializeMungers(config.PRMungersList, &config.Config)
 			return doMungers(config)
 		},
 	}
 	config.AddRootFlags(root)
 	addMungeFlags(config, root)
 
-	prMungers := pulls.GetAllMungers()
+	prMungers := mungers.GetAllMungers()
 	for _, m := range prMungers {
-		m.AddFlags(root, &config.Config)
-	}
-
-	issueMungers := issues.GetAllMungers()
-	for _, m := range issueMungers {
 		m.AddFlags(root, &config.Config)
 	}
