+++ mungegithub/mungers/ok-to-test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package pulls
+package mungers
 
 import (
 	github_util "k8s.io/contrib/mungegithub/github"
